+++ plat/rockchip/rk3399/drivers/soc/soc.c
@@ -34,6 +34,7 @@
 #include <mmio.h>
 #include <platform_def.h>
 #include <plat_private.h>
+#include <dram.h>
 #include <rk3399_def.h>
 #include <rk3399m0.h>
 #include <soc.h>
@@ -410,4 +411,5 @@ void plat_rockchip_soc_init(void)
 	soc_global_soft_reset_init();
 	plat_rockchip_gpio_init();
 	soc_m0_init();
+	dram_init();
 }
